guides/language-identification/requirements.txt
numpy==1.23.5
datasets==3.1.0
torch==2.4.1
torchaudio==2.4.1
soundfile==0.13.1
librosa==0.9.2
espnet==202412
espnet-model-zoo==0.1.7
speechbrain==1.0.2
matplotlib==3.10.3
langcodes==3.5.0
pandas==2.2.3
scikit-learn==1.6.1
scipy==1.15.3
xgboost==3.0.2
